+++ tests/time_utils/human_readable_time_parsing_test.py
@@ -1,25 +1,31 @@
 import datetime
 
+import pytest
 import pytz
-
 from freezegun import freeze_time
-
 from trailscraper.time_utils import parse_human_readable_time
+from tzlocal import get_localzone
+
+
+def test_shoud_parse_into_datetime_with_timezone_information():
+    assert parse_human_readable_time("now").tzinfo is not None
 
 
 def test_should_parse_full_dates():
-    assert parse_human_readable_time("2017-12-22") == \
+    assert parse_human_readable_time("2017-12-22").replace(tzinfo=None) == \
            datetime.datetime(2017, 12, 22, 0, 0, 0)
 
 
 def test_should_parse_full_datetimes():
-    assert parse_human_readable_time("2017-12-22 10:11:12") == \
+    assert parse_human_readable_time("2017-12-22 10:11:12").replace(tzinfo=None) == \
            datetime.datetime(2017, 12, 22, 10, 11, 12)
+
 
 @freeze_time("2010-11-12 13:14:15")
 def test_should_parse_human_readable_current_time():
     assert parse_human_readable_time("now").astimezone(pytz.utc) == \
            datetime.datetime(2010,11,12,13,14,15,tzinfo=pytz.utc)
+
 
 @freeze_time("2010-11-12 13:14:15")
 def test_should_parse_human_readable_relative_times():